+++ cmd/influxd/launcher/launcher.go
@@ -448,7 +448,7 @@ func (m *Launcher) run(ctx context.Context) (err error) {
 		// https://www.vaultproject.io/docs/commands/index.html#environment-variables
 		svc, err := vault.NewSecretService()
 		if err != nil {
-			m.logger.Error("failed initalizing vault secret service", zap.Error(err))
+			m.logger.Error("failed initializing vault secret service", zap.Error(err))
 			return err
 		}
 		secretSvc = svc
@@ -614,7 +614,11 @@ func (m *Launcher) run(ctx context.Context) (err error) {
 		LookupService:                   lookupSvc,
 		DocumentService:                 m.kvService,
 		OrgLookupService:                m.kvService,
+		WriteEventRecorder:              infprom.NewEventRecorder("write"),
+		QueryEventRecorder:              infprom.NewEventRecorder("query"),
 	}
+
+	m.reg.MustRegister(m.apibackend.PrometheusCollectors()...)
 
 	// HTTP server
 	httpLogger := m.logger.With(zap.String("service", "http"))